# factory/contracts/openapi/factory-control.v1.json
{
  "openapi": "3.1.0",
  "info": {"title": "Adaptive Factory Local Control API", "version": "1.0.0"},
  "paths": {
    "/health/live": {"get": {"tags": ["health"], "responses": {"200": {"description": "live"}}}},
    "/health/ready": {"get": {"tags": ["health"], "responses": {"200": {"description": "ready"}}}},
    "/v1/tasks": {"get": {"description": "Requires task:list scope.", "tags": ["tasks"], "security": [{"bearerAuth": []}], "responses": {"200": {"description": "bounded list"}}}, "post": {"description": "Requires task:submit scope.", "tags": ["tasks"], "security": [{"bearerAuth": []}], "responses": {"200": {"description": "duplicate"}, "201": {"description": "created"}}}},
    "/v1/tasks/{task_id}": {"get": {"description": "Requires task:read scope.", "tags": ["tasks"], "security": [{"bearerAuth": []}], "responses": {"200": {"description": "task"}, "404": {"description": "not found"}}}},
    "/v1/tasks/{task_id}/cancel": {"post": {"description": "Requires task:cancel scope.", "tags": ["tasks"], "security": [{"bearerAuth": []}], "responses": {"200": {"description": "cancelled"}}}},
    "/v1/claims": {"post": {"description": "Requires task:claim scope.", "tags": ["worker"], "security": [{"bearerAuth": []}], "responses": {"200": {"description": "claim result"}}}},
    "/v1/heartbeats": {"post": {"description": "Requires task:heartbeat scope.", "tags": ["worker"], "security": [{"bearerAuth": []}], "responses": {"200": {"description": "lease"}}}},
    "/v1/proposals": {"post": {"description": "Requires task:release scope.", "tags": ["worker"], "security": [{"bearerAuth": []}], "responses": {"200": {"description": "release status"}}}},
    "/v1/budget-reservations": {"post": {"description": "Requires task:budget scope and a current actor-bound lease.", "tags": ["worker"], "security": [{"bearerAuth": []}], "responses": {"200": {"description": "idempotent reservation identity"}}}},
    "/v1/usage-observations": {"post": {"description": "Requires task:budget scope and settles current-run reservations.", "tags": ["worker"], "security": [{"bearerAuth": []}], "responses": {"200": {"description": "idempotent usage observation"}}}},
    "/v1/kill-switches": {"post": {"description": "Requires factory:kill scope.", "tags": ["operator"], "security": [{"bearerAuth": []}], "responses": {"200": {"description": "kill state"}}}},
    "/v1/reconcile": {"post": {"description": "Requires factory:reconcile scope.", "tags": ["operator"], "security": [{"bearerAuth": []}], "responses": {"200": {"description": "bounded repair result"}}}},
    "/metrics": {"get": {"description": "Requires operator factory:reconcile scope; returns bounded low-cardinality JSON counters.", "tags": ["operator"], "security": [{"bearerAuth": []}], "responses": {"200": {"description": "redacted metric families"}}}}
  },
  "components": {"securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer"}}}
}